frontend/requirements.txt
pandas==2.3.0
plotly==6.0.1
pymysql==1.1.1
python-dotenv==1.1.1
Requests==2.32.4
streamlit==1.46.1
